LayoutTests/storage/indexeddb/modern/idbobjectstore-clear-1-expected.txt

This test creates an object store then populates it. It then clears it and makes sure it has nothing left in it.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


ALERT: Initial upgrade needed: Old version - 0 New version - 1
ALERT: Initial upgrade versionchange transaction complete
ALERT: Success opening database connection - Starting readwrite transaction
ALERT: Value gotten was bar1
ALERT: Value gotten was bar2
ALERT: Value gotten was bar3
ALERT: Value gotten was bar4
ALERT: Value gotten was bar5
ALERT: Value gotten was bar6
ALERT: Value gotten was bar7
ALERT: Value gotten was bar8
ALERT: Value gotten was bar9
ALERT: Object store cleared
ALERT: Value gotten was undefined
ALERT: Value gotten was undefined
ALERT: Value gotten was undefined
ALERT: Value gotten was undefined
ALERT: Value gotten was undefined
ALERT: Value gotten was undefined
ALERT: Value gotten was undefined
ALERT: Value gotten was undefined
ALERT: Value gotten was undefined
ALERT: Readwrite transaction complete
PASS successfullyParsed is true

TEST COMPLETE
